Added NLs workaround
