+++ src/metatensor/models/alchemical_model/tests/test_invariance.py
@@ -6,12 +6,11 @@ import torch
 from metatensor.torch.atomistic import (
     ModelCapabilities,
     ModelOutput,
-    NeighborsListOptions,
 )
 
 from metatensor.models.alchemical_model import DEFAULT_HYPERS, Model
 
-from ..utils import get_rascaline_neighbors_list
+from metatensor.models.utils.neighbors_lists import get_rascaline_neighbors_list
 from . import DATASET_PATH
 
 
@@ -29,19 +28,19 @@ def test_rotational_invariance():
         },
     )
     alchemical_model = Model(capabilities, DEFAULT_HYPERS["model"]).to(torch.float64)
+    requested_neighbors_lists = alchemical_model.requested_neighbors_lists()
 
     structure = ase.io.read(DATASET_PATH)
     original_structure = copy.deepcopy(structure)
     structure.rotate(48, "y")
-
-    nl_options = NeighborsListOptions(model_cutoff=5.0, full_list=True)
     original_system = rascaline.torch.systems_to_torch(original_structure)
-    nl = get_rascaline_neighbors_list(original_system, nl_options)
-    original_system.add_neighbors_list(nl_options, nl)
-
     system = rascaline.torch.systems_to_torch(structure)
-    nl = get_rascaline_neighbors_list(system, nl_options)
-    system.add_neighbors_list(nl_options, nl)
+    for nl_options in requested_neighbors_lists:
+        nl = get_rascaline_neighbors_list(original_system, nl_options)
+        original_system.add_neighbors_list(nl_options, nl)
+
+        nl = get_rascaline_neighbors_list(system, nl_options)
+        system.add_neighbors_list(nl_options, nl)
 
     original_output = alchemical_model(
         [original_system],
