+++ src/metatensor/models/alchemical_model/utils/__init__.py
@@ -1,9 +1,5 @@
 from .systems_to_torch_spex_dict import systems_to_torch_spex_dict
-from .get_ase_neighbors_list import get_ase_neighbors_list
-from .get_rascaline_neighbors_list import get_rascaline_neighbors_list
 
 __all__ = [
     "systems_to_torch_spex_dict",
-    "get_ase_neighbors_list",
-    "get_rascaline_neighbors_list",
 ]
