+++ src/metatensor/models/cli/eval_model.py
@@ -13,6 +13,7 @@ from ..utils.info import finalize_aggregated_info, update_aggregated_info
 from ..utils.loss import TensorMapDictLoss
 from ..utils.model_io import load_model
 from ..utils.omegaconf import expand_dataset_config
+from ..utils.neighbors_lists import check_and_update_neighbors_lists
 from .formatter import CustomHelpFormatter
 
 
@@ -72,6 +73,11 @@ def _eval_targets(model, dataset: Union[_BaseDataset, torch.utils.data.Subset]) 
             value_or_gradient: 0.0 for value_or_gradient in value_or_gradient_list
         }
     loss_fn = TensorMapDictLoss(loss_weights_dict)
+
+    # Check and update the neighborlists:
+    if hasattr(model, "requested_neighbors_lists"):
+        requested_neighbors_lists = model.requested_neighbors_lists()
+        check_and_update_neighbors_lists(dataset, requested_neighbors_lists)
 
     # Create a dataloader:
     dataloader = torch.utils.data.DataLoader(
